+++ tests/06-writing-doubles/Reference.cpp
@@ -11,7 +11,7 @@ int main() {
     output.precision(17);
     for (int i = 2; i < n; ++i) {
         v2 = a * v0 + b * v1 + c;
-        output << ' ' << (v2 < 0 ? -sqrt(-v2) : sqrt(v2));
+        output << ' ' << (v2 < 0 ? -sqrt((double) (-v2)) : sqrt((double) (v2)));
         v0 = v1;
         v1 = v2;
     }
