+++ tests/06-writing-doubles/Test.cpp
@@ -8,7 +8,7 @@ int main() {
     io << v0 << ' ' << v1;
     for (int i = 2; i < n; ++i) {
         v2 = a * v0 + b * v1 + c;
-        io << ' ' << (v2 < 0 ? -sqrt(-v2) : sqrt(v2));
+        io << ' ' << (v2 < 0 ? -sqrt((double) (-v2)) : sqrt((double) (v2)));
         v0 = v1;
         v1 = v2;
     }
